% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ptarmigan.R
\docType{data}
\name{Ptarmigan}
\alias{Ptarmigan}
\title{Ptarmigan data set}
\format{An object of class \code{data.frame} with 89 rows and 16 columns.}
\usage{
data(Ptarmigan)
}
\description{
Example data set
}
\keyword{datasets}
